refactor: rename MediaCoverCard to MediaItemCard
