+++ projects/client/src/lib/sections/lists/components/MediaItemCard.svelte
